add v3 tests
Contrary to rust-protobuf, test_basics are different from v2/v3 because
proto files ARE different.

I tried to replicate the correct behavior as much as possible

There is one test in particular that doesn't work:
- must not write if default values, as per proto3 methodology
