+++ tests/rust_protobuf/mod.rs
@@ -10,4 +10,4 @@ mod test_utils;
 mod v2;
 
 //#[cfg(proto3)]
-//mod v3;
+mod v3;
