+++ tests/rust_protobuf/v3/mod.rs
@@ -1,14 +1,15 @@
-extern crate protobuf;
-
 mod test_basic_pb;
 mod test_basic;
 
-mod test_zeros_are_not_written;
+// mod test_zeros_are_not_written;
 
 mod test_map_pb;
 mod test_map;
 
-mod test_enum_alias_pb;
-mod test_enum_alias;
+// mod test_enum_alias_pb;
+// mod test_enum_alias;
 
 mod test_ident_pb;
+
+mod test_oneof_pb;
+mod test_oneof;
