+++ tests/rust_protobuf/v3/test_map.rs
@@ -1,48 +1,48 @@
-use protobuf::text_format::print_to_string;
+use quick_protobuf::*;
 
+use std::borrow::Cow;
+use rust_protobuf::hex::{encode_hex, decode_hex};
 use super::test_map_pb::*;
-
-use test::*;
 
 #[test]
 fn test_map() {
-    let mut map = TestMap::new();
-    let mut entry = TestMapEntry::new();
-    entry.set_v(10);
+    let mut map = TestMap::default();
+    let mut entry = TestMapEntry::default();
+    entry.v = Some(10);
 
-    test_serialize_deserialize("", &map);
+    test_serialize_deserialize!("", &map, TestMap);
 
-    map.mut_m().insert("two".to_owned(), 2);
-    test_serialize_deserialize("0a 07 0a 03 74 77 6f 10 02", &map);
+    map.m.insert(Cow::Borrowed("two"), 2);
+    test_serialize_deserialize!("0a 07 0a 03 74 77 6f 10 02", &map, TestMap);
 
-    map.mut_m().insert("sixty six".to_owned(), 66);
+    map.m.insert(Cow::Borrowed("sixty six"), 66);
     // Insert map entry sub message
-    map.mut_mm().insert("map".to_owned(), entry);
+    map.mm.insert(Cow::Borrowed("map"), entry);
     // cannot (easily) test hex, because order is not specified
-    test_serialize_deserialize_no_hex(&map);
+    test_serialize_deserialize_length_delimited!(&map, TestMap);
 }
 
 #[test]
 fn test_map_with_object() {
-    let mut map = TestMap::new();
+    let mut map = TestMap::default();
 
-    let mut entry = TestMapEntry::new();
-    entry.set_v(10);
+    let mut entry = TestMapEntry::default();
+    entry.v = Some(10);
 
-    test_serialize_deserialize("", &map);
+    test_serialize_deserialize!("", &map, TestMap);
 
-    map.mut_mm().insert("map".to_owned(), entry);
+    map.mm.insert(Cow::Borrowed("map"), entry);
     // cannot (easily) test hex, because order is not specified
-    test_serialize_deserialize_no_hex(&map);
+    test_serialize_deserialize_length_delimited!(&map, TestMap);
 }
 
-#[test]
-fn text_format() {
-    let mut map = TestMap::new();
-
-    assert_eq!(&*print_to_string(&map), "");
-
-    map.mut_m().insert("two".to_owned(), 2);
-
-    assert_eq!(&*print_to_string(&map), "m {key: \"two\" value: 2}")
-}
+// #[test]
+// fn text_format() {
+//     let mut map = TestMap::new();
+// 
+//     assert_eq!(&*print_to_string(&map), "");
+// 
+//     map.mut_m().insert("two".to_owned(), 2);
+// 
+//     assert_eq!(&*print_to_string(&map), "m {key: \"two\" value: 2}")
+// }
